api/stock_api/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.0
pandas>=2.0
numpy>=1.24
akshare>=1.12
python-dotenv>=1.0
